+++ Services/Order/Order.Host/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -48,8 +48,8 @@ namespace Order.Host.Migrations
                     b.Property<long>("Quantity")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("TotalPrice")
-                        .HasColumnType("bigint");
+                    b.Property<decimal>("TotalPrice")
+                        .HasColumnType("numeric");
 
                     b.Property<string>("UserId")
                         .IsRequired()
